+++ src/bitstamp.rs
@@ -75,7 +75,7 @@ fn symbol_to_string(symbol: &Symbol) -> String {
 
 pub async fn run_bitstamp(
     summary_tx: UnboundedSender<Summary>,
-    symbol: &Symbol, depth: u16,
+    symbol: &Symbol, depth: usize,
 ) -> Result<(), Box<dyn std::error::Error>> {
     let tracer = global::tracer("run_bitstamp");
     let span = tracer.start("running bitstamp");
